+++ packages/auth0-acul-js/interfaces/export/common.ts
@@ -10,3 +10,4 @@ export type { Organizations } from '../models/user';
 export type { PasswordValidationResult } from '../utils/validate-password';
 export type { Identifier, IdentifierType as ScreenIdentifierType } from '../utils/get-enabled-identifiers';
 export type { UsernameValidationResult, UsernameValidationError } from '../utils/validate-username';
+export type { MfaPushPollingControl, MfaPushPollingError, MfaPushPollingCompleteCallback, MfaPushPollingErrorCallback, MfaPushPollingOptions } from '../common/mfa-push-polling';
