+++ packages/auth0-acul-js/interfaces/screens/reset-password-mfa-push-challenge-push.ts
@@ -1,7 +1,7 @@
 import type { CustomOptions } from '../common';
+import type { MfaPushPollingControl, MfaPushPollingOptions } from '../common/mfa-push-polling';
 import type { BaseMembers } from '../models/base-context';
 import type { ScreenMembers } from '../models/screen';
-import type { MfaPushPollingError, MfaPushPollingControl } from '../screens/mfa-push-challenge-push';
 
 /**
  * Interface for the screen data specific to reset-password-mfa-push-challenge-push screen
@@ -46,8 +46,7 @@ export interface ResetPasswordMfaPushChallengePushMembers extends BaseMembers {
 
   /**
    * Allows polling for the push notification challenge to be approved.
-   * @param intervalMs Polling interval in milliseconds
-   * @param onComplete Callback function to be called when polling is completed
+   * @param options Optional polling configuration options
    */
-  pollingManager(intervalMs: number, onComplete: () => void, onError?: (error: MfaPushPollingError) => void): MfaPushPollingControl;
+  pollingManager(options?: MfaPushPollingOptions): MfaPushPollingControl;
 }
