+++ packages/auth0-acul-js/src/screens/signup/index.ts
@@ -9,9 +9,7 @@ import { validateUsername as _validateUsername} from '../../utils/validate-usern
 import { ScreenOverride } from './screen-override';
 import { TransactionOverride } from './transaction-override';
 
-import type { Identifier } from '../../../interfaces/utils/get-enabled-identifiers';
 import type { ScreenContext } from '../../../interfaces/models/screen';
-import type { UsernameValidationResult } from '../../../interfaces/utils/validate-username';
 import type { TransactionContext } from '../../../interfaces/models/transaction';
 import type {
   SignupMembers,
@@ -21,6 +19,8 @@ import type {
   TransactionMembersOnSignup as TransactionOptions,
 } from '../../../interfaces/screens/signup';
 import type { FormOptions } from '../../../interfaces/utils/form-handler';
+import type { Identifier } from '../../../interfaces/utils/get-enabled-identifiers';
+import type { UsernameValidationResult } from '../../../interfaces/utils/validate-username';
 
 export default class Signup extends BaseContext implements SignupMembers {
   static screenIdentifier: string = ScreenIds.SIGNUP;
